requirements.txt
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
psycopg2-binary
apache-airflow
